azure-kusto-data/pyproject.toml
[project]
name = "azure-kusto-data"
version = "5.0.0"
description = "Kusto Data Client"
authors = [
    { name = "Microsoft Corporation", email = "kustalk@microsoft.com" },
]
license = "MIT"
readme = "README.rst"
urls = { "Repository" = "https://github.com/Azure/azure-kusto-python" }

dependencies = [
    "azure-core>=1.30.0",
    "azure-identity>=1.15.0",
    "ijson>=3.2.3",
    "msal>=1.27.0",
    "python-dateutil>=2.8.2",
    "requests>=2.31.0",
]
requires-python = ">= 3.8"

[project.optional-dependencies]
pandas = [
    "pandas>=2.0.3",
]
async = [
    "aiohttp>=3.9.3",
    "asgiref>=3.7.2",
]

[tool.uv]
package = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/azure"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
